+++ include/utils.h
@@ -25,16 +25,20 @@ limitations under the License.
 
 #include <pybind11/pybind11.h>
 
-#include <exception>   // std::rethrow_exception, std::current_exception
-#include <functional>  // std::hash
-#include <sstream>     // std::ostringstream
-#include <string>      // std::string
-#include <utility>     // std::move, std::pair, std::make_pair
-#include <vector>      // std::vector
+#include <exception>      // std::rethrow_exception, std::current_exception
+#include <functional>     // std::hash
+#include <sstream>        // std::ostringstream
+#include <string>         // std::string
+#include <unordered_map>  // std::unordered_map
+#include <utility>        // std::move, std::pair, std::make_pair
+#include <vector>         // std::vector
 
 namespace py = pybind11;
 using size_t = py::size_t;
 using ssize_t = py::ssize_t;
+
+// The maximum size of the type cache.
+constexpr ssize_t MAX_TYPE_CACHE_SIZE = 4096;
 
 // boost::hash_combine
 template <class T>
@@ -49,6 +53,58 @@ inline void HashCombine(py::ssize_t& seed, const T& v) {  // NOLINT[runtime/refe
     // NOLINTNEXTLINE[cppcoreguidelines-avoid-magic-numbers]
     seed ^= (hasher(v) + 0x9e3779b9 + (seed << 6) + (seed >> 2));
 }
+
+class TypeHash {
+ public:
+    using is_transparent = void;
+    py::size_t operator()(const py::object& t) const { return std::hash<PyObject*>{}(t.ptr()); }
+    py::size_t operator()(const py::handle& t) const { return std::hash<PyObject*>{}(t.ptr()); }
+};
+class TypeEq {
+ public:
+    using is_transparent = void;
+    bool operator()(const py::object& a, const py::object& b) const { return a.ptr() == b.ptr(); }
+    bool operator()(const py::object& a, const py::handle& b) const { return a.ptr() == b.ptr(); }
+    bool operator()(const py::handle& a, const py::object& b) const { return a.ptr() == b.ptr(); }
+    bool operator()(const py::handle& a, const py::handle& b) const { return a.ptr() == b.ptr(); }
+};
+
+class NamedTypeHash {
+ public:
+    using is_transparent = void;
+    py::size_t operator()(const std::pair<std::string, py::object>& p) const {
+        py::size_t seed = 0;
+        HashCombine(seed, p.first);
+        HashCombine(seed, p.second.ptr());
+        return seed;
+    }
+    py::size_t operator()(const std::pair<std::string, py::handle>& p) const {
+        py::size_t seed = 0;
+        HashCombine(seed, p.first);
+        HashCombine(seed, p.second.ptr());
+        return seed;
+    }
+};
+class NamedTypeEq {
+ public:
+    using is_transparent = void;
+    bool operator()(const std::pair<std::string, py::object>& a,
+                    const std::pair<std::string, py::object>& b) const {
+        return a.first == b.first && a.second.ptr() == b.second.ptr();
+    }
+    bool operator()(const std::pair<std::string, py::object>& a,
+                    const std::pair<std::string, py::handle>& b) const {
+        return a.first == b.first && a.second.ptr() == b.second.ptr();
+    }
+    bool operator()(const std::pair<std::string, py::handle>& a,
+                    const std::pair<std::string, py::object>& b) const {
+        return a.first == b.first && a.second.ptr() == b.second.ptr();
+    }
+    bool operator()(const std::pair<std::string, py::handle>& a,
+                    const std::pair<std::string, py::handle>& b) const {
+        return a.first == b.first && a.second.ptr() == b.second.ptr();
+    }
+};
 
 constexpr bool NONE_IS_LEAF = true;
 constexpr bool NONE_IS_NODE = false;
@@ -399,7 +455,25 @@ inline bool IsNamedTupleClassImpl(const py::handle& type) {
     return false;
 }
 inline bool IsNamedTupleClass(const py::handle& type) {
-    return PyType_Check(type.ptr()) && IsNamedTupleClassImpl(type);
+    if (!PyType_Check(type.ptr())) [[unlikely]] {
+        return false;
+    }
+
+    static auto cache = std::unordered_map<py::handle, bool, TypeHash, TypeEq>{};
+    auto it = cache.find(type);
+    if (it != cache.end()) [[likely]] {
+        return it->second;
+    }
+    bool result = IsNamedTupleClassImpl(type);
+    if (cache.size() < MAX_TYPE_CACHE_SIZE) [[likely]] {
+        cache.emplace(type, result);
+        (void)py::weakref(type, py::cpp_function([type](py::handle weakref) -> void {
+                              cache.erase(type);
+                              weakref.dec_ref();
+                          }))
+            .release();
+    }
+    return result;
 }
 inline bool IsNamedTupleInstance(const py::handle& object) {
     return IsNamedTupleClass(py::type::handle_of(object));
@@ -462,7 +536,25 @@ inline bool IsStructSequenceClassImpl(const py::handle& type) {
     return false;
 }
 inline bool IsStructSequenceClass(const py::handle& type) {
-    return PyType_Check(type.ptr()) && IsStructSequenceClassImpl(type);
+    if (!PyType_Check(type.ptr())) [[unlikely]] {
+        return false;
+    }
+
+    static auto cache = std::unordered_map<py::handle, bool, TypeHash, TypeEq>{};
+    auto it = cache.find(type);
+    if (it != cache.end()) [[likely]] {
+        return it->second;
+    }
+    bool result = IsStructSequenceClassImpl(type);
+    if (cache.size() < MAX_TYPE_CACHE_SIZE) [[likely]] {
+        cache.emplace(type, result);
+        (void)py::weakref(type, py::cpp_function([type](py::handle weakref) -> void {
+                              cache.erase(type);
+                              weakref.dec_ref();
+                          }))
+            .release();
+    }
+    return result;
 }
 inline bool IsStructSequenceInstance(const py::handle& object) {
     return IsStructSequenceClass(py::type::handle_of(object));
@@ -476,6 +568,16 @@ inline void AssertExactStructSequence(const py::handle& object) {
         throw py::value_error("Expected an instance of PyStructSequence type, got " +
                               PyRepr(object) + ".");
     }
+}
+inline py::tuple StructSequenceGetFieldsImpl(const py::handle& type) {
+    const auto n_sequence_fields = getattr(type, Py_Get_ID(n_sequence_fields)).cast<ssize_t>();
+    auto* members = reinterpret_cast<PyTypeObject*>(type.ptr())->tp_members;
+    py::tuple fields{n_sequence_fields};
+    for (ssize_t i = 0; i < n_sequence_fields; ++i) {
+        // NOLINTNEXTLINE[cppcoreguidelines-pro-bounds-pointer-arithmetic]
+        SET_ITEM<py::tuple>(fields, i, py::str(members[i].name));
+    }
+    return fields;
 }
 inline py::tuple StructSequenceGetFields(const py::handle& object) {
     py::handle type;
@@ -492,12 +594,24 @@ inline py::tuple StructSequenceGetFields(const py::handle& object) {
         }
     }
 
-    const auto n_sequence_fields = getattr(type, Py_Get_ID(n_sequence_fields)).cast<ssize_t>();
-    auto* members = reinterpret_cast<PyTypeObject*>(type.ptr())->tp_members;
-    py::tuple fields{n_sequence_fields};
-    for (ssize_t i = 0; i < n_sequence_fields; ++i) {
-        // NOLINTNEXTLINE[cppcoreguidelines-pro-bounds-pointer-arithmetic]
-        SET_ITEM<py::tuple>(fields, i, py::str(members[i].name));
+    static auto cache = std::unordered_map<py::handle, py::tuple, TypeHash, TypeEq>{};
+    auto it = cache.find(type);
+    if (it != cache.end()) [[likely]] {
+        return it->second;
+    }
+    py::tuple fields = StructSequenceGetFieldsImpl(type);
+    if (cache.size() < MAX_TYPE_CACHE_SIZE) [[likely]] {
+        cache.emplace(type, fields);
+        fields.inc_ref();
+        (void)py::weakref(type, py::cpp_function([type](py::handle weakref) -> void {
+                              auto it = cache.find(type);
+                              if (it != cache.end()) [[likely]] {
+                                  it->second.dec_ref();
+                                  cache.erase(it);
+                              }
+                              weakref.dec_ref();
+                          }))
+            .release();
     }
     return fields;
 }