+++ tests/test_typing.py
@@ -18,11 +18,13 @@
 import re
 import sys
 import time
+import weakref
+from collections import namedtuple
 
 import pytest
 
 import optree
-from helpers import CustomNamedTupleSubclass, CustomTuple, Vector2D
+from helpers import CustomNamedTupleSubclass, CustomTuple, Vector2D, getrefcount
 
 
 class FakeNamedTuple(tuple):
@@ -82,6 +84,127 @@ def test_is_namedtuple():
     assert not optree.is_namedtuple_class(FakeStructSequence)
 
 
+def test_is_namedtuple_cache():
+    Point = namedtuple('Point', ('x', 'y'))  # noqa: PYI024
+
+    refcount = getrefcount(Point)
+    weakrefcount = weakref.getweakrefcount(Point)
+    assert optree.is_namedtuple(Point)
+    new_refcount = getrefcount(Point)
+    new_weakrefcount = weakref.getweakrefcount(Point)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    assert optree.is_namedtuple_class(Point)
+    assert weakref.getweakrefcount(Point) == new_weakrefcount
+    wr = weakref.getweakrefs(Point)[0]
+    assert wr() is Point
+    del Point
+    getrefcount()
+    assert wr() is None
+
+    refcount = getrefcount(time.struct_time)
+    weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert not optree.is_namedtuple(time.struct_time)
+    new_refcount = getrefcount(time.struct_time)
+    new_weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert new_refcount == refcount
+    assert new_weakrefcount <= weakrefcount + 1
+    assert not optree.is_namedtuple_class(time.struct_time)
+    assert weakref.getweakrefcount(time.struct_time) == new_weakrefcount
+
+    called_with = ''
+
+    class FooMeta(type):
+        def __del__(cls):
+            nonlocal called_with
+            called_with = cls.__name__
+
+    class Foo(metaclass=FooMeta):
+        pass
+
+    refcount = getrefcount(Foo)
+    weakrefcount = weakref.getweakrefcount(Foo)
+    assert not optree.is_namedtuple(Foo)
+    new_refcount = getrefcount(Foo)
+    new_weakrefcount = weakref.getweakrefcount(Foo)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    assert not optree.is_namedtuple_class(Foo)
+    assert weakref.getweakrefcount(Foo) == new_weakrefcount
+
+    assert called_with == ''
+    wr = weakref.getweakrefs(Foo)[0]
+    assert wr() is Foo
+    del Foo
+    getrefcount()
+    assert called_with == 'Foo'
+    assert wr() is None
+
+
+def test_namedtuple_fields_cache():
+    Point = namedtuple('Point', ('x', 'y'))  # noqa: PYI024
+
+    refcount = getrefcount(Point)
+    weakrefcount = weakref.getweakrefcount(Point)
+    assert optree.namedtuple_fields(Point) == ('x', 'y')
+    new_refcount = getrefcount(Point)
+    new_weakrefcount = weakref.getweakrefcount(Point)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    assert optree.namedtuple_fields(Point(0, 1)) == ('x', 'y')
+    assert weakref.getweakrefcount(Point) == new_weakrefcount
+    wr = weakref.getweakrefs(Point)[0]
+    assert wr() is Point
+
+    fields = optree.namedtuple_fields(Point)
+    assert optree.namedtuple_fields(Point) is fields
+    assert optree.namedtuple_fields(Point(0, 1)) is fields
+    new_fields = ('a', 'b')
+    Point._fields = new_fields
+    assert optree.namedtuple_fields(Point) is new_fields
+    assert optree.namedtuple_fields(Point(0, 1)) is new_fields
+
+    del Point
+    getrefcount()
+    assert wr() is None
+
+    with pytest.raises(
+        TypeError,
+        match=r"Expected a collections.namedtuple type, got <class '.*'>\.",
+    ):
+        assert optree.namedtuple_fields(time.struct_time)
+
+    called_with = ''
+
+    class FooMeta(type):
+        def __del__(cls):
+            nonlocal called_with
+            called_with = cls.__name__
+
+    class Foo(metaclass=FooMeta):
+        pass
+
+    refcount = getrefcount(Foo)
+    weakrefcount = weakref.getweakrefcount(Foo)
+    with pytest.raises(
+        TypeError,
+        match=r"Expected a collections.namedtuple type, got <class '.*'>\.",
+    ):
+        optree.namedtuple_fields(Foo)
+    new_refcount = getrefcount(Foo)
+    new_weakrefcount = weakref.getweakrefcount(Foo)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+
+    assert called_with == ''
+    wr = weakref.getweakrefs(Foo)[0]
+    assert wr() is Foo
+    del Foo
+    getrefcount()
+    assert called_with == 'Foo'
+    assert wr() is None
+
+
 def test_is_structseq():
     with pytest.raises(TypeError, match="type 'structseq' is not an acceptable base type"):
 
@@ -116,6 +239,63 @@ def test_is_structseq():
     assert not optree.is_structseq_class(FakeNamedTuple)
     assert not optree.is_structseq_class(Vector2D)
     assert not optree.is_structseq_class(FakeStructSequence)
+
+
+def test_is_structseq_cache():
+    Point = namedtuple('Point', ('x', 'y'))  # noqa: PYI024
+
+    refcount = getrefcount(Point)
+    weakrefcount = weakref.getweakrefcount(Point)
+    assert not optree.is_structseq(Point)
+    new_refcount = getrefcount(Point)
+    new_weakrefcount = weakref.getweakrefcount(Point)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    assert not optree.is_structseq_class(Point)
+    assert weakref.getweakrefcount(Point) == new_weakrefcount
+    wr = weakref.getweakrefs(Point)[0]
+    assert wr() is Point
+    del Point
+    getrefcount()
+    assert wr() is None
+
+    refcount = getrefcount(time.struct_time)
+    weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert optree.is_structseq(time.struct_time)
+    new_refcount = getrefcount(time.struct_time)
+    new_weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert new_refcount == refcount
+    assert new_weakrefcount <= weakrefcount + 1
+    assert optree.is_structseq_class(time.struct_time)
+    assert weakref.getweakrefcount(time.struct_time) == new_weakrefcount
+
+    called_with = ''
+
+    class FooMeta(type):
+        def __del__(cls):
+            nonlocal called_with
+            called_with = cls.__name__
+
+    class Foo(metaclass=FooMeta):
+        pass
+
+    refcount = getrefcount(Foo)
+    weakrefcount = weakref.getweakrefcount(Foo)
+    assert not optree.is_structseq(Foo)
+    new_refcount = getrefcount(Foo)
+    new_weakrefcount = weakref.getweakrefcount(Foo)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    assert not optree.is_structseq_class(Foo)
+    assert weakref.getweakrefcount(Foo) == new_weakrefcount
+
+    assert called_with == ''
+    wr = weakref.getweakrefs(Foo)[0]
+    assert wr() is Foo
+    del Foo
+    getrefcount()
+    assert called_with == 'Foo'
+    assert wr() is None
 
 
 def test_namedtuple_fields():
@@ -270,3 +450,64 @@ def test_structseq_fields():
         match=r"Expected a PyStructSequence type, got <class '.*\.FakeStructSequence'>\.",
     ):
         optree.structseq_fields(FakeStructSequence)
+
+
+def test_structseq_fields_cache():
+    Point = namedtuple('Point', ('x', 'y'))  # noqa: PYI024
+
+    refcount = getrefcount(Point)
+    weakrefcount = weakref.getweakrefcount(Point)
+    with pytest.raises(TypeError, match=r"Expected a PyStructSequence type, got <class '.*'>\."):
+        optree.structseq_fields(Point)
+    new_refcount = getrefcount(Point)
+    new_weakrefcount = weakref.getweakrefcount(Point)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+    with pytest.raises(
+        TypeError,
+        match=re.escape('Expected an instance of PyStructSequence type, got Point(x=0, y=1).'),
+    ):
+        optree.structseq_fields(Point(0, 1))
+    assert weakref.getweakrefcount(Point) == new_weakrefcount
+    wr = weakref.getweakrefs(Point)[0]
+    assert wr() is Point
+    del Point
+    getrefcount()
+    assert wr() is None
+
+    refcount = getrefcount(time.struct_time)
+    weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert optree.structseq_fields(time.struct_time) is optree.structseq_fields(time.struct_time)
+    new_refcount = getrefcount(time.struct_time)
+    new_weakrefcount = weakref.getweakrefcount(time.struct_time)
+    assert new_refcount == refcount
+    assert new_weakrefcount <= weakrefcount + 2
+    assert optree.structseq_fields(time.gmtime()) is optree.structseq_fields(time.struct_time)
+    assert weakref.getweakrefcount(time.struct_time) == new_weakrefcount
+
+    called_with = ''
+
+    class FooMeta(type):
+        def __del__(cls):
+            nonlocal called_with
+            called_with = cls.__name__
+
+    class Foo(metaclass=FooMeta):
+        pass
+
+    refcount = getrefcount(Foo)
+    weakrefcount = weakref.getweakrefcount(Foo)
+    with pytest.raises(TypeError, match=r"Expected a PyStructSequence type, got <class '.*'>\."):
+        optree.structseq_fields(Foo)
+    new_refcount = getrefcount(Foo)
+    new_weakrefcount = weakref.getweakrefcount(Foo)
+    assert new_refcount == refcount
+    assert new_weakrefcount == weakrefcount + 1
+
+    assert called_with == ''
+    wr = weakref.getweakrefs(Foo)[0]
+    assert wr() is Foo
+    del Foo
+    getrefcount()
+    assert called_with == 'Foo'
+    assert wr() is None